backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv

pydantic
pydantic-settings

SQLAlchemy
psycopg[binary]
pgvector

llama-index
llama-index-vector-stores-postgres

google-generativeai
httpx
tenacity

pypdf